Move celery compression task to separate file
