+++ components/job-orchestration/job_orchestration/scheduler/compress/compression_scheduler.py
@@ -31,7 +31,7 @@ from clp_py_utils.compression import validate_path_and_get_info
 from clp_py_utils.core import read_yaml_config_file
 from clp_py_utils.s3_utils import s3_get_object_metadata
 from clp_py_utils.sql_adapter import SQL_Adapter
-from job_orchestration.executor.compress.compression_task import compress
+from job_orchestration.executor.compress.celery_compress import compress
 from job_orchestration.scheduler.compress.partition import PathsToCompressBuffer
 from job_orchestration.scheduler.constants import (
     CompressionJobStatus,
